refactoring: domain層の中に、インフラのインターフェースを定義する
