+++ api/domain/repository/model.go
@@ -1,0 +1,1 @@
+package repository
