+++ api/infra/persistence/shared.go
@@ -1,4 +1,4 @@
-package repository
+package persistence
 
 import "fmt"
