+++ api/infra/persistence/user.go
@@ -1,23 +1,17 @@
-package repository
+package persistence
 
 import (
 	"errors"
 
 	"github.com/htoyoda18/TweetAppV2/api/domain/model"
+	"github.com/htoyoda18/TweetAppV2/api/domain/repository"
 	"github.com/htoyoda18/TweetAppV2/api/shared"
 	"gorm.io/gorm"
 )
 
-type User interface {
-	Get(*model.User, *gorm.DB) (*model.User, error)
-	Add(*model.User, *gorm.DB) (*model.User, error)
-	UpdatePassword(*model.User, *gorm.DB) error
-	UpdateUser(*model.User, *gorm.DB) error
-}
-
 type user struct{}
 
-func NewUser() User {
+func NewUser() repository.User {
 	return user{}
 }
