+++ api/usecase/user.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/htoyoda18/TweetAppV2/api/controller/handler/request"
 	"github.com/htoyoda18/TweetAppV2/api/domain/model"
-	"github.com/htoyoda18/TweetAppV2/api/repository"
+	"github.com/htoyoda18/TweetAppV2/api/domain/repository"
 	"github.com/htoyoda18/TweetAppV2/api/shared"
 	"gorm.io/gorm"
 )
